==> 03_data_ana/01_read_field20.ipynb <==
from config import *

import re

# Files already have

1. non-destructive field measure
    1. 0515, 0522, 0526, 0528 -> pure field max length
    2. `ground.data_non.destructive.xlsx` -> id map + 0528 data
    
2. destructive field measure -> id map + 0528 data -> field min|max length + weight

root = r"Y:\hwang_Pro\data\2020_tanashi_broccoli\03_field_measurement"

sheet_name = ["22-May", "26-May", "28-May"]
dates = [20200522, 20200526, 20200528]

non_dest = pd.DataFrame(columns=["date", "Ground_ID", "UAV_ID", "flower_size_max"])

for sheet, date in zip(sheet_name, dates):
    non_link = pd.read_excel(f"{root}/ground.data_non.destructive.xlsx", sheet_name=sheet)
    non_link["date"] = date
    non_link = non_link.loc[:, ["date", "Ground_ID", "UAV_ID", "flower_size_max"]]
    non_dest = pd.concat([non_dest, non_link])
    
non_dest

non_dest.to_csv("out/non_destructive20.csv", index=False)

dest = pd.DataFrame(columns=["date", "Ground_ID", "UAV_ID", "flower_size_max", "flower_size_min", "weight"])

for sheet, date in zip(sheet_name, dates):
    link = pd.read_excel(f"{root}/ground.data_destructive.xlsx", sheet_name=sheet)
    link["date"] = date
    link = link.loc[:, ["date", "Ground_ID", "UAV_ID", "flower_size_max", "flower_size_min", "weight"]]
    dest = pd.concat([dest, link])
    
dest

dest = dest.dropna()
dest

dest.to_csv("out/destructive20.csv", index=False)
