==> 03_data_ana/01_read_field21.ipynb <==
from config import *

import re
import platform
import pandas as pd

# Four different Files:
1. ID map -> the link between QGIS shp id & field measure
2. Non-destructive test (West + East) -> field measure (id, 0 45 90 135)
3. Destructive -> field measure weight (very regular)

uav date: "210512", "210514", "210515", "210519", "210520", "210526

user = platform.node()

if user == "DESKTOP-3J8SGRC":
    root = r"Y:\hwang_Pro\data\2021_tanashi_broccoli\03_field_measurement"
if user == "RTX3090AMD":
    root = r"Z:\hwang_Pro\data\2021_tanashi_broccoli\03_field_measurement"
elif user == "NERV":
    root = "./input/field_21"

excel = {"root": root,
         "drop": ["x", "X", "根",""],
         "link": {
             "name":"ID Map.xlsx",
             "circle": 2,
             "sheet": [
                 {"name": "West side",
                  "st_row": 1,  # count from 0
                  "ed_row": 143, 
                  "usecol": "B:Y",},
                 {"name": "East side",
                  "st_row": 1,  # count from 0
                  "ed_row": 145,
                  "usecol": "D:AA"},
             ],
         },
         "non-e": {
             "name":"non-destructive test(in East side).xlsx",
             "circle": 5,
             "sheet": [
                 {"name": "20210512",
                  "st_row": 1,  # count from 0
                  "ed_row": 145,
                  "usecol": "F:J,O:S,X:AB,AG:AK"},
                 {"name": "20210515",
                  "st_row": 1,  # count from 0
                  "ed_row": 145,
                  "usecol": "F:J,O:S,X:AB,AG:AK"},
                 {"name": "20210517",
                  "st_row": 1,  # count from 0
                  "ed_row": 145,
                  "usecol": "F:J,O:S,X:AB,AG:AK"},
                 {"name": "20210519",
                  "st_row": 1,  # count from 0
                  "ed_row": 145,
                  "usecol": "F:J,O:S,X:AB,AG:AK"},
                 {"name": "20210520",
                  "st_row": 1,  # count from 0
                  "ed_row": 145,
                  "usecol": "F:J,O:S,X:AB,AG:AK"},
             ]
         },
         "non-w" :{
             "name":"non-destructive test(in West side).xlsx",
             "circle": 5,
             "sheet": [
                 {"name": "20210515",
                  "st_row": 1,  # count from 0
                  "ed_row": 143,
                  "usecol": "I:M,R:V,AB:AF,AI:AM"},
                 {"name": "20210517",
                  "st_row": 1,  # count from 0
                  "ed_row": 143,
                  "usecol": "I:M,R:V,Y:AC,AF:AJ"},
                 {"name": "20210519",
                  "st_row": 1,  # count from 0
                  "ed_row": 143,
                  "usecol": "I:M,R:V,AB:AF,AI:AM"},
                 {"name": "20210520",
                  "st_row": 1,  # count from 0
                  "ed_row": 143,
                  "usecol": "I:M,R:V,AB:AF,AI:AM"},
                 {"name": "20210524",
                  "st_row": 1,  # count from 0
                  "ed_row": 143,
                  "usecol": "I:M,R:V,AB:AF,AI:AM"},
                 {"name": "20210526",
                  "st_row": 1,  # count from 0
                  "ed_row": 143,
                  "usecol": "I:M,R:V,AB:AF,AI:AM"},
             ]
         },
         "des":{
             "name": "destructive test.xlsx",
             "circle": None,
             "sheet": [
                 {"name": "20210514-E(No.1)", 
                  "st_row": 1,
                  "usecol": "B:G"},
                 {"name": "20210514-W(No.1)", 
                  "st_row": 1,
                  "usecol": "B:G"},
                 {"name": "20210517-E(No.2)", 
                  "st_row": 1,
                  "usecol": "B:G"},
                 {"name": "20210517-W(No.2)", 
                  "st_row": 1,
                  "usecol": "B:G"},
                 {"name": "20210519-E(No.3)", 
                  "st_row": 1,
                  "usecol": "B:G"},
                 {"name": "20210519-W(No.3)", 
                  "st_row": 1,
                  "usecol": "B:G"},
                 {"name": "20210520-E(No.4)", 
                  "st_row": 1,
                  "usecol": "B:G"},
                 {"name": "20210520-W(No.4)",
                  "st_row": 1,
                  "usecol": "B:G"},
             ]
         }
        }

# Part 1: Get link id

def read_sheet(file, sheet_id, endrow=True):
    
    file_name = os.path.join(excel["root"], excel[file]["name"])
    if endrow:
        source = pd.read_excel(file_name, 
                              sheet_name=excel[file]["sheet"][sheet_id]["name"], 
                              header=excel[file]["sheet"][sheet_id]["st_row"],
                              usecols=excel[file]["sheet"][sheet_id]["usecol"],
                              nrows=excel[file]["sheet"][sheet_id]["ed_row"],
                              dtype=str)
    else:
        source = pd.read_excel(file_name, 
                              sheet_name=excel[file]["sheet"][sheet_id]["name"], 
                              header=excel[file]["sheet"][sheet_id]["st_row"],
                              usecols=excel[file]["sheet"][sheet_id]["usecol"],
                              dtype=str)
        
    circle = excel[file]["circle"]
    if circle is not None:
        col_name = source.columns.values.tolist()
        
        st = 0
        ed = st+circle
        
        fix_name = col_name[st:ed]
        
        merge = source.iloc[:,st:ed]
        while ed < len(col_name):
            st += circle
            ed += circle
            
            to_merge = source.iloc[:,st:ed]
            
            rename_dict = {i:j for i,j in zip(to_merge.columns.values.tolist(), fix_name)}
            
            merge = pd.concat([merge, to_merge.rename(columns=rename_dict)])
        
        merge = merge.dropna()
        return merge.reset_index(drop=True)
            
    else:
        return source

link_w = read_sheet("link", 0)
link_w = link_w.rename(columns={"No.": "field", "欠：x": "img"})
print(f"remove\n{link_w[link_w.img.str.contains('[a-zA-Z]')]}")
link_w = link_w[~link_w.img.str.contains("[a-zA-Z]")]
link_w = link_w.reset_index(drop=True)
link_w["pos"] = "West"
link_w

link_e = read_sheet("link", 1)
link_e = link_e.rename(columns={"No.": "field", "ID": "img"})

print(f"remove\n{link_e[link_e.img.str.contains('[a-zA-Z]')]}")
link_e = link_e[~link_e.img.str.contains("[a-zA-Z]")]
link_e = link_e.reset_index(drop=True)
link_e["pos"] = "East"
link_e

link = pd.concat([link_w, link_e])
link = link.astype({"field":str, "pos": str, "img": float})
link = link.astype({"field":str, "pos": str, "img": int})
link

link.to_csv("out/link_table.csv", index=False)

# Part 2: Get Non-destructive

def clean_data(df, rename_dict, find_col, dtype_dict, insert_col=None):
    print(df.columns.values)
    df = df.rename(columns=rename_dict)
    idx = df[find_col].str.contains('[a-zA-Z]')
    print(f"remove\n{df[idx]}")
    df = df[~idx]
    df = df.reset_index(drop=True)
    
    if insert_col is not None:
        for k, v in insert_col.items():
            df[k] = v
    
    return df

non_e = read_sheet("non-e", 0)
non_e = clean_data(non_e, {"No..1": "field"}, find_col=90, 
                   dtype_dict={"field":str, 0:float, 90:float, 45:float, 135:float},
                   insert_col={"date": "20210506", "pos": "East"})
non_e

non_list = []

for i, sheet_e in enumerate(excel["non-e"]["sheet"]):
    date = sheet_e["name"]
    
    non_e = read_sheet("non-e", i)
    non_e = clean_data(non_e, {"No..1": "field"}, find_col=90, 
                       dtype_dict={"field":str, 0:float, 90:float, 45:float, 135:float},
                       insert_col={"date": date, "pos": "East"})
    non_list.append(non_e)
    
    
for i, sheet_w in enumerate(excel["non-w"]["sheet"]):
    date = sheet_w["name"]
    non_w = read_sheet("non-w", i)
    non_w = clean_data(non_w, {"No..2": "field", '0.1':0, '90.1':90, '45.1':45, '135.1':135}, find_col=90, 
                       dtype_dict={"field":str, 0:float, 90:float, 45:float, 135:float},
                       insert_col={"date": date, "pos": "West"})
    non_list.append(non_w)
    
non = pd.concat(non_list)
non

non = non[["date", "pos", "field", 0, 90, 45, 135]]
non = non.sort_values(by=['date', 'field'])
non

non.to_csv("out/non_destructive.csv", index=False)

# Part 3: Get Destructive

des_e = read_sheet("des", 0, endrow=False)
des_e = clean_data(des_e, {"Unnamed: 1": "field", "0° (mm)":0, "90° (mm)":90, "45° (mm)":45, "135° (mm)":135, "Weight (g)": "weight_g"}, 
                   find_col=0, 
                   dtype_dict={"field":str, 0:int, 90:int, 45:int, 135:int, "weight_g":float},
                   insert_col={"date": "20210506", "pos": "East"})
des_e

des_list = []

pos_dict = {'W':"West", "E":"East"}

for i, sheet_d in enumerate(excel["des"]["sheet"]):
    sheet_name = sheet_d["name"]
    date, pos, _ = re.split(f"\-|\(", sheet_name)

    des = read_sheet("des", i, endrow=False)
    des = clean_data(des, {"Unnamed: 1": "field", "0° (mm)":0, "90° (mm)":90, "45° (mm)":45, "135° (mm)":135, "Weight (g)": "weight_g"}, 
                     find_col=0, 
                     dtype_dict={"field":str, 0:int, 90:int, 45:int, 135:int, "weight_g":float},
                     insert_col={"date": date, "pos": pos_dict[pos]})

    des_list.append(des)
    
des = pd.concat(des_list)
des

des = des[["date", "pos", "field", 0, 90, 45, 135, "weight_g"]]
des = des.sort_values(by=['date', 'field'])
des

des = des.drop_duplicates()
des

des.to_csv("out/destructive_220827.csv", index=False)
