==> 03_data_ana/30_des_nondes_id.ipynb <==
import pandas as pd
import numpy as np

non_des_2020 = pd.read_excel("out/2020_measured_fix.xlsx")

non_des_2021 = pd.read_csv("out/non_destructive_manual.csv")
non_des_2021_all = pd.read_excel("out/2021_measured_fix.xlsx")

des_2020 = pd.read_csv("out/destructive20.csv")
des_2021 = pd.read_csv("out/destructive21.csv")
des_2021_link = pd.read_csv("out/link_table.csv")

# non des 2020

non_des_2020

non_des_2020_id = np.unique(non_des_2020['label'].to_numpy())
non_des_2020_id

len(non_des_2020_id)

print(non_des_2020_id.tolist())

# non des 2021

non_des_2021

non_des_2021_link_dict = des_2021_link.set_index('field')['img'].to_dict()
non_des_2021_link_dict

non_des_2021_id = []
for k in non_des_2021.field.drop_duplicates().values:
    if k in non_des_2021_link_dict.keys():
        non_des_2021_id.append(non_des_2021_link_dict[k])
non_des_2021_id

len(non_des_2021_id)

print(non_des_2021_id)

# des 2020

des_2020_id = des_2020['UAV_ID'].astype(int).to_list()

print(des_2020_id)

5916 seems not correct

len(des_2020_id)

# des 2021

des_2021['field']

des_2021_id = []
for k in des_2021['field'].unique():
    if k in non_des_2021_link_dict.keys():
        des_2021_id.append(non_des_2021_link_dict[k])
des_2021_id

len(des_2021_id)

print(des_2021_id)
