==> 03_data_ana/99_get_bbox_size.ipynb <==
import numpy as np
import pandas as pd
import shapefile

shp_path = "box_anno_0329.shp"

shp = shapefile.Reader(shp_path)

for i, shape in enumerate(shp.shapes()):
    coord_np = np.asarray(shape.points)

    x0, y0 = coord_np[0,0], coord_np[0,1]
    x1, y1 = coord_np[1,0], coord_np[1,1]
    x2, y2 = coord_np[2,0], coord_np[2,1]

    l1 = np.sqrt( (x0 - x1)**2 + (y0 - y1)**2 )
    l2 = np.sqrt( (x1 - x2)**2 + (y1 - y2)**2 )

    idx = shp.records()[i][0]
    print(idx, l1, l2)
    break

out = pd.DataFrame(columns=["id", "long(cm)", "short(cm)"])

for i, shape in enumerate(shp.shapes()):
    coord_np = np.asarray(shape.points)

    x0, y0 = coord_np[0,0], coord_np[0,1]
    x1, y1 = coord_np[1,0], coord_np[1,1]
    x2, y2 = coord_np[2,0], coord_np[2,1]

    l1 = np.sqrt( (x0 - x1)**2 + (y0 - y1)**2 )
    l2 = np.sqrt( (x1 - x2)**2 + (y1 - y2)**2 )

    idx = shp.records()[i][0]

    out.loc[len(out), :] = [idx, max(l1, l2)*100, min(l1, l2)*100]

out

out.to_csv("box_anno_0319.csv", index=False)
